+++ ChargeBee/Models/Order.cs
@@ -1269,6 +1269,10 @@ namespace ChargeBee.Models
                 get { return GetEnum<ValidationStatusEnum>("validation_status", false); }
             }
 
+            public int Index {
+                get { return GetValue<int>("index", true); }
+            }
+
         }
         public class OrderBillingAddress : Resource
         {