+++ ChargeBee/Models/Transaction.cs
@@ -264,6 +264,10 @@ namespace ChargeBee.Models
         {
             get { return GetValue<string>("business_entity_id", false); }
         }
+        public string PaymentMethodDetails 
+        {
+            get { return GetValue<string>("payment_method_details", false); }
+        }
         
         #endregion